Removed old networking from web2; fixed distrib pom
