+++ kura/org.eclipse.kura.web2/src/main/resources/org/eclipse/kura/web/client/messages/Messages.properties
@@ -569,7 +569,6 @@ netHwLAC=LAC
 netHwCI=CI
 netCurrentHwWifiChannel=Current WLAN Channel
 
-netStatusWarning=There is another interface configured as WAN. That interface should bet set to LAN before changing the selected interface to WAN. Multiple WAN interfaces is not recommended.
 netRouter=DHCPv4 & NAT
 netRouterMode=Router Mode
 netRouterDhcpBeginningAddress=DHCP Beginning Address
