+++ kura/org.eclipse.kura.web2/src/main/resources/org/eclipse/kura/web/client/messages/Messages_ja.properties
@@ -547,7 +547,6 @@ netHwUSBDevice=USBデバイス
 netHwSignalStrength=受信したシグナル強度（dBm）
 netCurrentHwWifiChannel=現在の Wlan チャネル
 
-netStatusWarning=WANとして環境設定された他のインターフェースが存在します。洗濯されたインターフェースをWANにする前に、そちらをLANに変更することが推奨されます。複数のインターフェースをWAN設定にすることは推奨されていません。
 netRouter=DHCPv4とNAT
 netRouterMode=ルータモード
 netRouterDhcpBeginningAddress=DHCP開始アドレス
